requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
databricks-sql-connector[pyarrow]==3.1.1
et_xmlfile==1.1.0
Faker==24.4.0
idna==3.6
joblib==1.3.2
lz4==4.3.3
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.1
psutil==5.9.8
pybreaker==1.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
thrift==0.16.0
tzdata==2024.1
urllib3==2.2.1

# --- Sprint 3 AI & Agent Dependencies ---
google-genai==0.3.0
langchain==0.1.13
langchain-core==0.1.33
langchain-classic==0.0.1
langchain-google-genai==1.0.1
langchain-community==0.0.29
chromadb==0.4.24
sentence-transformers==2.5.1
langchain-huggingface==0.0.2